pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shuttle-stack"
version = "0.1.0"
description = "Fixed-route shuttle ADAS research foundation"
requires-python = ">=3.10"
dependencies = [
  "pycapnp>=2.0.0",
  "pyzmq>=26.0.0",
  "PyYAML>=6.0.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
shuttle-stack = "shuttle.cli:app"
robot-core = "robot_core.cli:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["shuttle*", "robot_core*"]
